+++ NEWS.md
@@ -6,6 +6,7 @@
 - Support word level details comparison (optional) (TODO - add link before release)
 - Improve options cancel, reset, and detecting changes requiring data reloading (TODO - add link before release) 
 - Add configuration option descriptions (TODO - add link before release)
+- Improve RTF reading to handle some invalid RTF contents (TODO - add link before release)
 
 ## Bug fixes
 - Fix issue with comment exporting with multiple folder loads (TODO - add link before release)